+++ container_transform/tests/converter_tests.py
@@ -24,3 +24,21 @@ class ConverterTests(TestCase):
         self.assertIsInstance(fig_output, str)
 
         self.assertEqual(0, len(conv.messages))
+
+    def test_compose_converter_out(self):
+        filename = './container_transform/tests/task.json'
+        conv = Converter(filename, 'ecs', 'compose')
+
+        fig_output = conv.convert()
+
+        self.assertIsInstance(fig_output, str)
+
+        self.assertEqual(0, len(conv.messages))
+
+    def test_compose_converter_in(self):
+        filename = './container_transform/tests/docker-compose.yml'
+        conv = Converter(filename, 'compose', 'ecs')
+
+        output = conv.convert()
+
+        self.assertIsInstance(output, str)